Wireless Glove Remove Controlled Robot Car
As a basic principle, communicating two processors with radio frequency. The coordinate data from the gloves (Arduino Nano) are sent to the car (Arduino Uno) with two nRF24L01 modules in the glove and robot car. This data from the glove is re-compiled to guide the car after it is sent to the car. As a result, the 2-axis (y and x-axis) direction movements made with the glove activate the car in the same direction.